src/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-functions-durable>=1.0.0b6
pandas
PyYAML 
gspread
sklearn
xgboost
azureml
environs
azureml-core
azureml-sdk
pathlib
requests
# azureml-train-automl
azureml-explain-model
numpy==1.19.3
mysql-connector-python
astroid==2.4.2
attrs==20.2.0
Automat==20.2.0
cffi==1.14.3
colorama==0.4.3
constantly==15.1.0
crochet==1.12.0
cryptography==3.1.1
cssselect==1.1.0
emacs==24.1.50.1.post20120610
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
isort==4.3.21
itemadapter==0.1.1
itemloaders==1.0.3
jmespath==0.10.0
lazy-object-proxy==1.4.3
lxml==4.5.2
mccabe==0.6.1
parsel==1.6.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pylint==2.5.3
pyodbc==4.0.30
pyOpenSSL==19.1.0
queuelib==1.5.0
Scrapy==2.3.0
service-identity==18.1.0
six==1.15.0
toml==0.10.1
Twisted==20.3.0
typed-ast==1.4.1
w3lib==1.22.0
wrapt==1.12.1
zope.interface==5.1.2
